chore(webpack): split configuration depending on the target
