+++ config/webpack.editor.js
@@ -1,11 +1,5 @@
-const LodashModuleReplacementPlugin = require("lodash-webpack-plugin");
-const CircularDependencyPlugin = require("circular-dependency-plugin");
-const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
-const webpack = require("webpack");
 const path = require("path");
-const WebpackNotifierPlugin = require("webpack-notifier");
 const packageName = require("../package.json").shortName;
-
 const relativeOutputPath = "build/umd";
 const finalPackageName = packageName + ".js";
 
@@ -28,7 +22,7 @@ const loaderSetup = {
   }
 };
 
-let config = {
+module.exports = {
   // TODO: check if targeting electron with webpack is better in some way
   target: "electron-main",
   // devtool is already set with -d (debug) and removed with -p (production) flags from webpack and webpack dev server
@@ -38,17 +32,8 @@ let config = {
   output: {
     filename: finalPackageName,
     path: path.resolve(relativeOutputPath),
-    // export itself to UMD format
-    libraryTarget: "umd",
-    umdNamedDefine: true,
     // webpack dev server hot reload path
     publicPath: relativeOutputPath
-  },
-  resolve: {
-    // Look for modules in .js files first
-    extensions: [".js", ".json"],
-    // Add 'src' to our modules, as all our app code will live in there, so Webpack should look in there for modules
-    modules: ["src", "node_modules"]
   },
   module: {
     rules: [
@@ -61,8 +46,5 @@ let config = {
         use: loaderSetup
       }
     ]
-  },
-  plugins: [new LodashModuleReplacementPlugin(), new WebpackNotifierPlugin({ alwaysNotify: true })]
+  }
 };
-
-module.exports = config;
