+++ demos/sdf-text/index.html
@@ -48,32 +48,65 @@
   <canvas id="canvas"></canvas>
 
   <div style="float:left; margin-left: 100px;">
-    <label>Text <textarea cols="40" rows="5" style="width:400px;" value="Lorem ipsum" id="str"></textarea></label><br>
-    <label>Letter Spacing <input type="number" id="letterSpacing" value="0"></label>
+    <label>Text
+      <textarea cols="40" rows="5" style="width:400px;" value="Lorem ipsum" id="str"></textarea>
+    </label>
+    <br>
+    <label>Letter Spacing
+      <input type="number" id="letterSpacing" value="0">
+    </label>
   </div>
 
   <div style="float:left;">
-    <label><input type="range" style="width:400px" value="70" step="0.01" min="6" max="128.0" id="scale">Size</label><br>
-    <label><input type="range" style="width:400px" value="0.0" step="1.0" min="0.0" max="10.0" id="stroke">Stroke Size</label><br>
-    <label><input type="range" style="width:400px" value="2" step="0.01" min="0" max="4" id="gamma">Gamma</label><br>
-    <label><input type="range" style="width:400px" value="5.0" step="1.0" min="0.0" max="10.0" id="dropShadow">DropShadow Smoothing</label><br>
-    <label><input type="range" style="width:400px" value="5.0" step="1.0" min="-10.0" max="10.0" id="dropShadowOffsetX">Shadow Offset X</label><br>
-    <label><input type="range" style="width:400px" value="5.0" step="1.0" min="-10.0" max="10.0" id="dropShadowOffsetY">Shadow Offset Y</label><br>
+    <label>
+      <input type="range" style="width:400px" value="70" step="0.01" min="6" max="128.0" id="scale">Size</label>
+    <br>
+    <label>
+      <input type="range" style="width:400px" value="0.0" step="1.0" min="0.0" max="10.0" id="stroke">Stroke Size</label>
+    <br>
+    <label>
+      <input type="range" style="width:400px" value="2" step="0.01" min="0" max="4" id="gamma">Gamma</label>
+    <br>
+    <label>
+      <input type="range" style="width:400px" value="5.0" step="1.0" min="0.0" max="10.0" id="dropShadow">DropShadow Smoothing</label>
+    <br>
+    <label>
+      <input type="range" style="width:400px" value="5.0" step="1.0" min="-10.0" max="10.0" id="dropShadowOffsetX">Shadow Offset X</label>
+    <br>
+    <label>
+      <input type="range" style="width:400px" value="5.0" step="1.0" min="-10.0" max="10.0" id="dropShadowOffsetY">Shadow Offset Y</label>
+    <br>
   </div>
 
   <div style="float:left;">
-    <label><input type="radio" name="group-1" id="alignLeft" value="left">Align Left</label><br>
-    <label><input type="radio" name="group-1" id="alignCenter" value="center">Align Center</label><br>
-    <label><input type="radio" name="group-1" id="alignRight" value="center">Align Right</label><br>
+    <label>
+      <input type="radio" name="group-1" id="alignLeft" value="left">Align Left</label>
+    <br>
+    <label>
+      <input type="radio" name="group-1" id="alignCenter" value="center">Align Center</label>
+    <br>
+    <label>
+      <input type="radio" name="group-1" id="alignRight" value="center">Align Right</label>
+    <br>
 
-    <label><input type="checkbox" id="wordwrap">Word Wrap</label><br>
-    <label><input type="checkbox" id="charwrap">Character Wrap</label><br>
+    <label>
+      <input type="checkbox" id="wordwrap">Word Wrap</label>
+    <br>
+    <label>
+      <input type="checkbox" id="charwrap">Character Wrap</label>
+    <br>
   </div>
 
   <div style="float:left;">
-    <label><input type="checkbox" id="debug">Debug Texture</label><br>
-    <label><input type="checkbox" id="outlineEnabled">Stroke</label><br>
-    <label><input type="checkbox" id="dropShadowEnabled">DropShadow</label><br>
+    <label>
+      <input type="checkbox" id="debug">Debug Texture</label>
+    <br>
+    <label>
+      <input type="checkbox" id="outlineEnabled">Stroke</label>
+    <br>
+    <label>
+      <input type="checkbox" id="dropShadowEnabled">DropShadow</label>
+    <br>
   </div>
 
   <!-- imports -->
